requirements.txt
aiogram==2.25.1
cython==3.0.0
vedis==0.7.2
telebot==0.0.5
Pillow==10.0.0
python-dotenv==1.0.0